+++ ProfessionalFullStackDeveloper/spring-boot-example/src/main/java/com/github/marcinciapa/tutorials/fullstack/customer/CustomerListDataAccessService.java
@@ -13,8 +13,8 @@ public class CustomerListDataAccessService implements CustomerDao {
 
     static {
         customers = new ArrayList<>();
-        Customer alex = new Customer(1, "Alex", "alex@gmail.com", 21);
-        Customer jamila = new Customer(2, "Jamila", "jamila@gmail.com", 19);
+        Customer alex = new Customer(1L, "Alex", "alex@gmail.com", 21);
+        Customer jamila = new Customer(2L, "Jamila", "jamila@gmail.com", 19);
         customers.add(alex);
         customers.add(jamila);
     }
@@ -25,7 +25,7 @@ public class CustomerListDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         return customers
                 .stream()
                 .filter(c -> id.equals(c.getId()))
@@ -44,7 +44,7 @@ public class CustomerListDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomer(Integer id) {
+    public void deleteCustomer(Long id) {
         customers.stream()
                 .filter(c -> !id.equals(c.getId()))
                 .findFirst()
@@ -52,7 +52,7 @@ public class CustomerListDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         return customers
                 .stream()
                 .anyMatch(c -> id.equals(c.getId()));
